+++ src/test/java/com/ajou_nice/with_pet/fixture/PetsitterApplicationDtoFixture.java
@@ -1,8 +1,8 @@
 package com.ajou_nice.with_pet.fixture;
 
 import com.ajou_nice.with_pet.domain.entity.User;
-import com.ajou_nice.with_pet.dto.applicant.PetsitterApplicationRequest;
-import com.ajou_nice.with_pet.dto.applicant.PetsitterApplicationResponse;
+import com.ajou_nice.with_pet.applicant.model.dto.PetsitterApplicationRequest;
+import com.ajou_nice.with_pet.applicant.model.dto.PetsitterApplicationResponse;
 import com.ajou_nice.with_pet.enums.Gender;
 
 import java.time.LocalDate;
